requirements/local.txt
-r base.txt

Werkzeug[watchdog]==2.3.7 # https://github.com/pallets/werkzeug
ipdb==0.13.13  # https://github.com/gotcha/ipdb
psycopg[c]==3.1.9  # https://github.com/psycopg/psycopg
watchfiles==0.20.0  # https://github.com/samuelcolvin/watchfiles

# Testing
# ------------------------------------------------------------------------------
mypy==1.4.1  # https://github.com/python/mypy
django-stubs[compatible-mypy]==4.2.3  # https://github.com/typeddjango/django-stubs
pytest==7.4.1  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.7  # https://github.com/Frozenball/pytest-sugar
djangorestframework-stubs[compatible-mypy]==3.14.2  # https://github.com/typeddjango/djangorestframework-stubs

# Documentation
# ------------------------------------------------------------------------------
sphinx==6.2.1  # https://github.com/sphinx-doc/sphinx
sphinx-autobuild==2021.3.14 # https://github.com/GaretJax/sphinx-autobuild

# Code quality
# ------------------------------------------------------------------------------
flake8==6.1.0  # https://github.com/PyCQA/flake8
flake8-isort==6.0.0  # https://github.com/gforcada/flake8-isort
coverage==7.3.0  # https://github.com/nedbat/coveragepy
black==23.7.0  # https://github.com/psf/black
djlint==1.32.1  # https://github.com/Riverside-Healthcare/djLint
pylint-django==2.5.3  # https://github.com/PyCQA/pylint-django
pylint-celery==0.3  # https://github.com/PyCQA/pylint-celery
pre-commit==3.4.0  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy==3.3.0  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==4.2.0  # https://github.com/jazzband/django-debug-toolbar
django-extensions==3.2.3  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==3.1.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==4.5.2  # https://github.com/pytest-dev/pytest-django

# Extra
djangocodemirror==2.1.0
